+++ src/App.tsx
@@ -1,7 +1,7 @@
-import { Routes, Route } from 'react-router-dom';
-import CartPage from './pages/cart/CartPage';
-import OrderSuccessPage from './pages/order/OrderSuccessPage';
-import { ROUTES } from './shared/config/routes';
+import { Routes, Route } from "react-router-dom";
+import CartPage from "./pages/cart/CartPage";
+import OrderSuccessPage from "./pages/order-success/OrderSuccessPage";
+import { ROUTES } from "./shared/config/routes";
 
 function App() {
   return (
